requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
confluent-kafka==2.12.2
Faker==38.0.0
idna==3.11
kafka==1.3.5
kafka-python==2.2.15
psycopg2-binary==2.9.11
pyspark==3.5.7
requests==2.32.5
tzdata==2025.2
urllib3==2.5.0
